+++ lib/bindings/fmvvmstatelesswidget.dart
@@ -34,7 +34,7 @@ abstract class FmvvmStatelessWidget<V extends BindableBase>
   @mustCallSuper
   Widget build(BuildContext context) {
     if (_isNavigable) {
-      Core.componentResolver.resolveType<NavigationService>().currentContext =
+      Core.componentResolver.resolveType<NavigationService>().viewContext =
           context;
     }
     return null;