+++ lib/bindings/navigationservice.dart
@@ -8,11 +8,26 @@ abstract class NavigationService {
   /// Navigates to a new viewmodel of the type specified by the generic.
   ///
   /// The [parameter] is a value that will be passed to the new viewmodel's
-  /// init method.
-  Future<void> navigate<T extends ViewModel>({Object parameter});
+  /// init method. The Future will be resolved with the ViewModel that is navigated to is
+  /// popped from the stack.
+  Future navigate<V extends ViewModel>({Object parameter});
+
+  /// Navigates to a new viewmodel of the type specified by the generic.
+  ///
+  /// The [parameter] is a value that will be passed to the new viewmodel's
+  /// init method. It is expexted that the ViewModel being navigated to will
+  /// return an instance of type R when it is popped off the stack.
+  Future<R> navigateForResult<R extends Object, V extends ViewModel>(
+      {Object parameter});
 
   /// Pops the current view / viewmodel off the stack and goes to the previous one.
   void navigateBack();
+
+  /// Pops the current view / viewmodel off the stack and goes to the previous one.
+  ///
+  /// The [paramter] is the result to send back to the calling view model.
+  /// This should be used in conjunction with navigateForResult.
+  void navigateBackWithResult<R extends Object>([R parameter]);
 
   /// Creates a view model of a specified type.
   ///
@@ -24,5 +39,5 @@ abstract class NavigationService {
 
   /// The current context. This method is called by default when a widget defined as a page
   /// has been shown.
-  set currentContext(BuildContext context);
+  set viewContext(BuildContext context);
 }