+++ lib/fmvvmnavigationservice.dart
@@ -2,24 +2,41 @@ part of fmvvm;
 
 /// Default implementation of the fmvvm NavigationService interface.
 class FmvvmNavigationService implements NavigationService {
-  BuildContext _currentContext;
+  BuildContext _viewContext;
 
   /// Navigates to a new viewmodel of the type specified by the generic.
   ///
   /// The [parameter] is a value that will be passed to the new viewmodel's
   /// init method.
-  Future<void> navigate<T extends ViewModel>({Object parameter}) async {
-    var _viewModel = createViewModel<T>(parameter);
+  Future navigate<V extends ViewModel>({Object parameter}) async {
+    var _viewModel = createViewModel<V>(parameter);
 
-    var routeName = Core._viewLocator.getViewFromViewModelType<T>();
+    var routeName = Core._viewLocator.getViewFromViewModelType<V>();
 
-    await Navigator.pushNamed(_currentContext, routeName,
-        arguments: _viewModel);
+    await Navigator.of(_viewContext)
+        .pushNamed(routeName, arguments: _viewModel);
+  }
+
+  @override
+  Future<R> navigateForResult<R extends Object, V extends ViewModel>(
+      {Object parameter}) async {
+    var _viewModel = createViewModel<V>(parameter);
+
+    var routeName = Core._viewLocator.getViewFromViewModelType<V>();
+
+    var returnValue = await Navigator.of(_viewContext)
+        .pushNamed<R>(routeName, arguments: _viewModel);
+
+    return returnValue;
   }
 
   /// Pops the current view / viewmodel off the stack and goes to the previous one.
   void navigateBack() {
-    Navigator.pop(_currentContext);
+    Navigator.of(_viewContext).pop();
+  }
+
+  void navigateBackWithResult<R extends Object>([R parameter]) {
+    Navigator.of(_viewContext).pop(parameter);
   }
 
   /// Creates a view model of a specified type.
@@ -36,7 +53,7 @@ class FmvvmNavigationService implements NavigationService {
 
   /// The current context. This method is called by default when a widget defined as a page
   /// has been shown.
-  set currentContext(BuildContext context) {
-    _currentContext = context;
+  set viewContext(BuildContext context) {
+    _viewContext = context;
   }
 }